requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.5
colorama==0.4.6
exceptiongroup==1.1.1
Faker==33.1.0
fdb==2.0.2
future==0.18.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
pycicd==23.364.734
pycparser==2.21
pyparsing==3.1.0
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.9.0.post0
SDK==1.0.0
selenium==4.10.0
six==1.16.0
sly==0.5
sniffio==1.3.0
sortedcontainers==2.4.0
tdb==2.23.424.720
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.12.2
urllib3==2.0.3
watchdog==3.0.0
wsproto==1.2.0
